pyproject.toml
[tool.poetry]
name = "warflower"
version = "0.1.0"
description = "Dedi server mgmt"
authors = ["scott"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
docker = "^7.0.0"
pyyaml = "^6.0.1"
flask = "^3.0.1"
flask-httpauth = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
